# packages/api-sdk/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users": {
      "post": {
        "operationId": "createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          }
        },
        "summary": "Create a new user",
        "tags": [
          "User"
        ]
      },
      "get": {
        "operationId": "listUsers",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Filter users by name or email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Search users",
        "tags": [
          "User"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "get one user by id",
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Delete user by id",
        "tags": [
          "User"
        ]
      }
    },
    "/api/products": {
      "post": {
        "operationId": "createProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductView"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Create a new product",
        "tags": [
          "Product"
        ]
      },
      "get": {
        "operationId": "listProducts",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Filter users by name or email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fridgeId",
            "required": false,
            "in": "query",
            "description": "Filter by given fridge",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fridgeLocation",
            "required": false,
            "in": "query",
            "description": "Filter by given location for fridges",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product(s) retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Search all products",
        "tags": [
          "Product"
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "get one product by id",
        "tags": [
          "Product"
        ]
      },
      "patch": {
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Update a product",
        "tags": [
          "Product"
        ]
      },
      "delete": {
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Delete product by id",
        "tags": [
          "Product"
        ]
      }
    },
    "/api/fridges": {
      "post": {
        "operationId": "createFridge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FridgeBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fridge created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FridgeView"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Create a new fridge",
        "tags": [
          "Fridge"
        ]
      },
      "get": {
        "operationId": "listFridges",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Filter users by name or email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fridges(s) retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FridgeView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Search all fridges",
        "tags": [
          "Fridge"
        ]
      }
    },
    "/api/fridges/{id}": {
      "get": {
        "operationId": "getFridge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "get one fridge by id",
        "tags": [
          "Fridge"
        ]
      },
      "patch": {
        "operationId": "updateFridge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FridgeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Update a fridge",
        "tags": [
          "Fridge"
        ]
      },
      "delete": {
        "operationId": "deleteFridge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Delete fridge by id",
        "tags": [
          "Fridge"
        ]
      }
    },
    "/api/recipes": {
      "post": {
        "operationId": "createRecipe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recipe created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeView"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Create a new recipe",
        "tags": [
          "Recipe"
        ]
      },
      "get": {
        "operationId": "listRecipes",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Filter users by name or email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recipes(s) retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipeView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Search all recipes",
        "tags": [
          "Recipe"
        ]
      }
    },
    "/api/recipes/{id}": {
      "get": {
        "operationId": "getRecipe",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "get one recipe by id",
        "tags": [
          "Recipe"
        ]
      },
      "patch": {
        "operationId": "updateRecipe",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Update a recipe",
        "tags": [
          "Recipe"
        ]
      },
      "delete": {
        "operationId": "deleteRecipe",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "x-auth": []
          }
        ],
        "summary": "Delete recipe by id",
        "tags": [
          "Recipe"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenView"
                }
              }
            }
          }
        },
        "summary": "log in",
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "info": {
    "title": "Node Course API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-auth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-auth"
      }
    },
    "schemas": {
      "UserBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "surname",
          "email",
          "password"
        ]
      },
      "UserView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "surname",
          "email"
        ]
      },
      "UpdateUserBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ProductBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "ownerId": {
            "type": "string",
            "nullable": true
          },
          "fridgeId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "size"
        ]
      },
      "ProductView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "ownerId": {
            "type": "string"
          },
          "fridgeId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "size"
        ]
      },
      "FridgeBody": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "floor": {
            "type": "number"
          },
          "capacity": {
            "type": "number"
          }
        },
        "required": [
          "address",
          "floor",
          "capacity"
        ]
      },
      "FridgeView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "address": {
            "type": "string"
          },
          "floor": {
            "type": "number"
          },
          "capacity": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "address",
          "floor",
          "capacity"
        ]
      },
      "RecipeBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "ownerId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "RecipeView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ]
      },
      "LoginBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AccessTokenView": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number"
          }
        },
        "required": [
          "token",
          "expiresIn"
        ]
      }
    }
  }
}